Change type of `manualScale` to `CGFloat`
This better reflects the duality between it an `UIScreen.scale` which
has type `CGFloat`

+++ FBSnapshotTestCase/FBSnapshotTestCase.h
@@ -125,7 +125,7 @@
 /**
  When set to values different than 0, it uses a specific scale for generating and comparing images rather than using the main screen's default scale.
  */
-@property (readwrite, nonatomic, assign) NSUInteger manualScale;
+@property (readwrite, nonatomic, assign) CGFloat manualScale;
 
 - (void)setUp NS_REQUIRES_SUPER;
 - (void)tearDown NS_REQUIRES_SUPER;
